+++ test/support/fake_pipedrive_server.ex
@@ -13,6 +13,7 @@ defmodule ExPipedrive.FakePipedriveServer do
   import ExPipedrive.FakeDealFieldApiHandler
   import ExPipedrive.FakeDealV2ApiHandler
   import ExPipedrive.FakeFieldV2ApiHandler
+  import ExPipedrive.FakeFilterApiHandler
   import ExPipedrive.FakeLeadApiHandler
   import ExPipedrive.FakeNoteApiHandler
   import ExPipedrive.FakeOrganizationApiHandler
@@ -210,6 +211,36 @@ defmodule ExPipedrive.FakePipedriveServer do
     conn
     |> put_resp_header("content-type", "application/json;charset=utf-8")
     |> handle_list_own_activities()
+  end
+
+  get "/api/v1/filters" do
+    conn
+    |> put_resp_header("content-type", "application/json;charset=utf-8")
+    |> handle_list_filters(conn.query_params)
+  end
+
+  post "/api/v1/filters" do
+    conn
+    |> put_resp_header("content-type", "application/json;charset=utf-8")
+    |> handle_create_filter()
+  end
+
+  put "/api/v1/filters/:id" do
+    conn
+    |> put_resp_header("content-type", "application/json;charset=utf-8")
+    |> handle_update_filter()
+  end
+
+  delete "/api/v1/filters/:id" do
+    conn
+    |> put_resp_header("content-type", "application/json;charset=utf-8")
+    |> handle_delete_filter(conn.params)
+  end
+
+  get "/api/v1/filters/:id" do
+    conn
+    |> put_resp_header("content-type", "application/json;charset=utf-8")
+    |> handle_get_filter(conn.params)
   end
 
   get "/api/v1/webhooks" do